code release v1

+++ blender_deform.py
@@ -1,9 +1,10 @@
+# DAWand: generate deformed primitives using Blender 
 from re import L
 import bpy
 import os
 import sys
 import numpy as np
-
+from pathlib import Path
 
 '''
 @input: 
@@ -13,7 +14,7 @@ import numpy as np
     deformed mesh
     
 to run it from cmd line:
-/opt/blender/blender --background --python blender_deform.py /home/rana/code/remesh/datasets/raw/cube.obj /home/rana/Downloads/temp.obj
+blender --background --python blender_deform.py -- --datadir ./datasets/meshgen/source_highres --exportdir ./datasets/deformed_primitives
 '''
 
 class Process:
@@ -39,7 +40,6 @@ class Process:
         mod.deform_axis = np.random.choice(['X', 'Y', 'Z'])
         mod.deform_method = np.random.choice(['TWIST', "BEND", "TAPER", "STRETCH"])
         
-        # For debugging
         print(f"Deform: {mod.deform_method}\nAxis: {mod.deform_axis}\nAngle: {mod.angle}\nFactor: {mod.factor}")
 
     def export_obj(self, mesh, export_name):
@@ -55,26 +55,47 @@ class Process:
                                  axis_forward='-Z', axis_up='Y')
 
 if __name__ == "__main__":  
-    datadir = "/Users/guanzhi/Documents/InteractiveSegmentation/MeshCNNFeatures/datasets/meshgen/source_highres"
-    outdir = "./datasets/meshgen/polyhedrons_def_highres"
-    import shutil 
-    for filename in os.listdir(outdir):
-        file_path = os.path.join(outdir, filename)
-        try:
-            if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
-        except Exception as e:
-            print('Failed to delete %s. Reason: %s' % (file_path, e))
-    n = 19
+    import sys 
+    
+    # Important: all named arguments should be passed after a double dash following the python script call e.g. '--python script.py -- -arg1 -arg2'
+    argv = sys.argv[sys.argv.index('--') + 1:]
+    
+    import argparse 
+    parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter)
+    parser.add_argument('--datadir', required=True, type=str, help='path to base primitives with ground truth segmentations in a "labels" folder')
+    parser.add_argument('--exportdir', required=True, type=str, help='export path')
+    parser.add_argument('--n', type=int, default=10, help='number of deformed meshes to generate per primitive')
+    parser.add_argument('--ndef_low', type=int, default=3, help='lower bound of deformations to apply')
+    parser.add_argument('--ndef_high', type=int, default=10, help='upper bound of deformations to apply')
+    parser.add_argument('--overwrite', action="store_true")
+    args = parser.parse_args(argv)
+    
+    n = args.n 
+    exportdir = args.exportdir 
+    datadir = args.datadir 
+    
+    if not os.path.exists(exportdir):
+        Path(exportdir).mkdir(exist_ok=True, parents=True)
+        
+    # Copy labels folder over 
+    # import shutil 
+    # shutil.copytree(os.path.join(datadir,"labels"), os.path.join(exportdir,"labels"))
+    
+    if args.overwrite:
+        for filename in os.listdir(exportdir):
+            file_path = os.path.join(exportdir, filename)
+            try:
+                if os.path.isfile(file_path) or os.path.islink(file_path):
+                    os.unlink(file_path)
+                elif os.path.isdir(file_path):
+                    shutil.rmtree(file_path)
+            except Exception as e:
+                print('Failed to delete %s. Reason: %s' % (file_path, e))
+
     for file in os.listdir(datadir):
         if not file.endswith(".obj"):
             continue 
         meshname = os.path.splitext(file)[0] 
-        # TODO: We start with simple shapes first 
-        if meshname in ['dodecahedron', 'icosahedron']:
-            continue
         for i in range(n):
-            n_def = np.random.choice(range(3,10))
-            Process(os.path.join(datadir, file), os.path.join(outdir, f"{meshname}{i}.obj"), n=n_def)
+            n_def = np.random.choice(range(args.ndef_low, args.ndef_high))
+            Process(os.path.join(datadir, file), os.path.join(exportdir, f"{meshname}{i}.obj"), n=n_def)
